serviço de deletar imagem

+++ src/services/dishDeleteService.js
@@ -1,4 +1,6 @@
+import knex from "../database/knex/index.js"
 import AppError from "../utils/AppError.js"
+import DiskStorage from "../providers/diskstorage.js"
 
 class DishDeleteService {
   constructor(dishRepository) {
@@ -6,12 +8,23 @@ class DishDeleteService {
   }
 
   async execute({ id }) {
+    const dish = await knex("dish").where({ id }).first()
+    if (!dish) {
+      throw new AppError("Informe um id vádido", 400)
+    }
     const dishDeleted = await this.dishRepository.delete({
       id,
     })
 
+    const diskStorage = new DiskStorage()
+
+    if (dish.image) {
+      await diskStorage.deleteFile(dish.image)
+    }
+
     return dishDeleted
   }
+  
 }
 
 export default DishDeleteService
